Use try/except clause when creating the .lcsim directory

+++ Workflow/Modules/LCSIMAnalysis.py
@@ -114,8 +114,13 @@ class LCSIMAnalysis(ModuleBase):
         
       aliasproperties = os.path.basename(self.aliasproperties)
       if os.path.exists(aliasproperties):
-        os.mkdir(os.path.join(homedir,".lcsim"))
-        shutil.copy(aliasproperties,os.path.join(homedir,".lcsim",aliasproperties))
+        try:
+          os.mkdir(os.path.join(homedir,".lcsim"))
+        except:
+          self.log.error("Could not create .lcsim in $HOME folder")
+        if os.path.exists(os.path.join(homedir,".lcsim")):
+          shutil.copy(aliasproperties,os.path.join(homedir,".lcsim",aliasproperties))
+
     
     #if tarfile.is_tarfile(self.sourcedir) :
     #  untarred_sourcedir = tarfile.open(self.sourcedir,'r')
